Transfer pcr primer design tests
